requirements.txt
annotated-types==0.6.0
anthropic==0.25.6
anyio==4.3.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.4
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.7
openai==1.21.2
packaging==24.0
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
retry==0.9.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tokenizers==0.19.1
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
